// src/interfaces/interfaces.ts
export interface Team {
  name: string | null;
  icon: string | null;
}

export interface Tournament {
  name: string | null;
  link: string | null;
}

export interface Match {
  teamLeft: Team;
  teamRight: Team;
  tournament: Tournament;
  dateTime: string | null;
  format: string | null;
  game: string | null;
}